cleanup(angular): move existing migrators into their own folder
