+++ packages/angular/src/generators/ng-add/migrators/e2e.migrator.ts
@@ -1,18 +1,20 @@
 import { cypressProjectGenerator } from '@nrwl/cypress';
 import { nxE2EPreset } from '@nrwl/cypress/plugins/cypress-preset';
 import { installedCypressVersion } from '@nrwl/cypress/src/utils/cypress-version';
+import type {
+  ProjectConfiguration,
+  TargetConfiguration,
+  Tree,
+} from '@nrwl/devkit';
 import {
   addProjectConfiguration,
   joinPathFragments,
   names,
   offsetFromRoot,
-  ProjectConfiguration,
   readJson,
   readProjectConfiguration,
   removeProjectConfiguration,
   stripIndents,
-  TargetConfiguration,
-  Tree,
   updateJson,
   updateProjectConfiguration,
   visitNotIgnoredFiles,
@@ -23,25 +25,27 @@ import { insertImport } from '@nrwl/workspace/src/utilities/ast-utils';
 import { getRootTsConfigPathInTree } from '@nrwl/workspace/src/utilities/typescript';
 import { tsquery } from '@phenomnomnominal/tsquery';
 import { basename, relative } from 'path';
+import type {
+  Node,
+  ObjectLiteralExpression,
+  PropertyAssignment,
+} from 'typescript';
 import {
   isObjectLiteralExpression,
   isPropertyAssignment,
   isStringLiteralLike,
   isTemplateExpression,
-  Node,
-  ObjectLiteralExpression,
-  PropertyAssignment,
   SyntaxKind,
 } from 'typescript';
-import { GeneratorOptions } from '../schema';
-import { FileChangeRecorder } from './file-change-recorder';
-import { Logger } from './logger';
-import { ProjectMigrator } from './project.migrator';
-import {
+import type { GeneratorOptions } from '../schema';
+import { FileChangeRecorder } from '../utilities/file-change-recorder';
+import type { Logger } from '../utilities/logger';
+import type {
   MigrationProjectConfiguration,
   Target,
   ValidationResult,
-} from './types';
+} from '../utilities/types';
+import { ProjectMigrator } from './project.migrator';
 
 type SupportedTargets = 'e2e';
 const supportedTargets: Record<SupportedTargets, Target> = {
