+++ packages/angular/src/generators/ng-add/migrators/index.ts
@@ -1,0 +1,4 @@
+export * from './app.migrator';
+export * from './e2e.migrator';
+export * from './lib.migrator';
+export * from './project.migrator';
