+++ packages/angular/src/generators/ng-add/ng-add.spec.ts
@@ -1,7 +1,7 @@
 import * as angularCliMigrator from './migrate-from-angular-cli';
 import * as initGenerator from '../init/init';
 import { ngAddGenerator } from './ng-add';
-import { Tree } from '@nrwl/devkit';
+import type { Tree } from '@nrwl/devkit';
 import { createTreeWithEmptyWorkspace } from '@nrwl/devkit/testing';
 
 describe('ngAdd generator', () => {
